+++ Feats/PsionicMeditation.cs
@@ -13,7 +13,7 @@ public static class PsionicMeditation
 {
     public static void Configure()
     {
-        var feat = FeatureConfigurator.New("PsionicMeditationFeat", Guids.PsionicMeditationFeat)
+        FeatureConfigurator.New("PsionicMeditationFeat", Guids.PsionicMeditationFeat)
             .SetDisplayName(Loc.Str("PW.PsionicMeditation.Name", "Psionic Meditation"))
             .SetDescription(Loc.Str("PW.PsionicMeditation.Desc", "You can take a move action to gain psionic focus, rather than a standard action."))
             .SetIcon(AbilityRefs.RestorationLesser.Reference.Get().Icon)