+++ Feats/PsionicWeapon.cs
@@ -3,6 +3,7 @@ using BlueprintCore.Actions.Builder;
 using BlueprintCore.Actions.Builder.ContextEx;
 using BlueprintCore.Blueprints.CustomConfigurators.Classes;
 using BlueprintCore.Blueprints.CustomConfigurators.Classes.Selection;
+using BlueprintCore.Blueprints.CustomConfigurators.UnitLogic.Buffs;
 using BlueprintCore.Blueprints.References;
 using BlueprintCore.Conditions.Builder;
 using BlueprintCore.Conditions.Builder.ContextEx;
@@ -14,6 +15,7 @@ using Kingmaker.Enums;
 using Kingmaker.Enums.Damage;
 using Kingmaker.RuleSystem;
 using Kingmaker.RuleSystem.Rules.Damage;
+using Kingmaker.UnitLogic.Buffs.Blueprints;
 using Kingmaker.UnitLogic.Mechanics;
 using PsychicWarrior.Utils;
 
@@ -23,13 +25,8 @@ public static class PsionicWeapon
 {
     public static void Configure()
     {
-        FeatureConfigurator.New("PsionicWeaponFeat", Guids.PsionicWeaponFeat)
-            .SetDisplayName(Loc.Str("PW.PsionicWeapon.Name", "Psionic Weapon"))
-            .SetDescription(Loc.Str("PW.PsionicWeapon.Desc",
-                "While psionically focused, your melee attacks deal additional force damage scaling with manifester level: 1d6 at ML 1, 2d6 at ML 6, 3d6 at ML 11, 4d6 at ML 16."))
-            .SetIcon(AbilityRefs.MagicWeaponGreater.Reference.Get().Icon)
-            .SetGroups(FeatureGroup.CombatFeat, FeatureGroup.Feat)
-            .AddPrerequisiteFeature(Guids.GainPsionicFocusFeature)
+        BuffConfigurator.New("PsionicWeaponBuff", Guids.PsionicWeaponBuff)
+            .SetFlags(BlueprintBuff.Flags.HiddenInUi)
             .AddContextRankConfig(ContextRankConfigs.CasterLevel().WithCustomProgression((5, 1), (10, 2), (15, 3), (20, 4)))
             .AddInitiatorAttackWithWeaponTrigger(
                 action: ActionsBuilder.New().Conditional(
@@ -41,6 +38,16 @@ public static class PsionicWeapon
                 onlyHit: true,
                 checkWeaponRangeType: true,
                 rangeType: WeaponRangeType.Melee)
+            .Configure();
+
+        FeatureConfigurator.New("PsionicWeaponFeat", Guids.PsionicWeaponFeat)
+            .SetDisplayName(Loc.Str("PW.PsionicWeapon.Name", "Psionic Weapon"))
+            .SetDescription(Loc.Str("PW.PsionicWeapon.Desc",
+                "While psionically focused, your melee attacks deal additional force damage scaling with manifester level: 1d6 at ML 1, 2d6 at ML 6, 3d6 at ML 11, 4d6 at ML 16."))
+            .SetIcon(AbilityRefs.MagicWeaponGreater.Reference.Get().Icon)
+            .SetGroups(FeatureGroup.CombatFeat, FeatureGroup.Feat)
+            .AddPrerequisiteFeature(Guids.GainPsionicFocusFeature)
+            .AddFacts([Guids.PsionicWeaponBuff])
             .AddRecommendedClass(Guids.PsychicWarriorClass)
             .Configure();